fix: isolate tempfile creation to platform specific file for tests
