+++ plugins/inputs/logfile/logfile_test.go
@@ -9,14 +9,12 @@ import (
 	"log"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strconv"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/aws/amazon-cloudwatch-agent/logs"
-	"github.com/aws/amazon-cloudwatch-agent/plugins/inputs/logfile/tail/winfile"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/text/encoding/simplifiedchinese"
@@ -1064,19 +1062,4 @@ func TestGenerateLogGroupName(t *testing.T) {
 		"The log group name %s is not the same as %s.",
 		logGroupName,
 		expectLogGroup))
-}
-
-func createTempFile(dir, prefix string) (*os.File, error) {
-	file, err := ioutil.TempFile(dir, prefix)
-	if runtime.GOOS == "windows" { // winfile open with FILE_SHARE_DELETE, allows the log file be deleted
-		if err := file.Close(); err != nil {
-			return nil, fmt.Errorf("Failed to close created temp file %v: %w", file.Name(), err)
-		}
-
-		file, err = winfile.OpenFile(file.Name(), os.O_RDWR, 0)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to open temp file for writing %v: %w", file.Name(), err)
-		}
-	}
-	return file, err
 }
