+++ plugins/inputs/logfile/tailersrc_test.go
@@ -9,33 +9,20 @@ import (
 	"io/ioutil"
 	"os"
 	"regexp"
-	"runtime"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/aws/amazon-cloudwatch-agent/logs"
 	"github.com/aws/amazon-cloudwatch-agent/plugins/inputs/logfile/tail"
-	"github.com/aws/amazon-cloudwatch-agent/plugins/inputs/logfile/tail/winfile"
 )
 
 func TestTailerSrc(t *testing.T) {
 
-	file, err := ioutil.TempFile("", "tailsrctest-*.log")
+	file, err := createTempFile("", "tailsrctest-*.log")
 	defer os.Remove(file.Name())
 	if err != nil {
 		t.Errorf("Failed to create temp file: %v", err)
-	}
-
-	if runtime.GOOS == "windows" { // winfile open with FILE_SHARE_DELETE, allows the log file be deleted
-		if err := file.Close(); err != nil {
-			t.Errorf("Failed to close created temp file %v: %v", file.Name(), err)
-		}
-
-		file, err = winfile.OpenFile(file.Name(), os.O_RDWR, 0)
-		if err != nil {
-			t.Errorf("Failed to open temp file for writing %v: %v", file.Name(), err)
-		}
 	}
 
 	statefile, err := ioutil.TempFile("", "tailsrctest-state-*.log")
