+++ plugins/inputs/logfile/tmpfile.go
@@ -1,0 +1,12 @@
+// +build !windows
+
+package logfile
+
+import (
+	"io/ioutil"
+	"os"
+)
+
+func createTempFile(dir, prefix string) (*os.File, error) {
+	return ioutil.TempFile(dir, prefix)
+}
